added 'add payload box', few more habitat changes

+++ rtty_dev/src/ukhas/HabitatRxEvent.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface HabitatRxEvent extends EventListener
 {
-	void HabitatRx(List<String> data, boolean success, String callsign, int startTime, int endTime);
+	void HabitatRx(List<String> data, boolean success, String callsign, long startTime, long endTime);
 }
